% Generated by roxygen2 (4.0.2): do not edit by hand
\name{owin2gpc}
\alias{gpc2owin}
\alias{owin2gpc}
\title{Conversion from and to the \code{"gpc.poly"} Class}
\usage{
owin2gpc(object)

gpc2owin(object, ...)
}
\arguments{
\item{object}{an object of class \code{"gpc.poly"} or \code{"owin"},
respectively.}

\item{...}{further arguments passed to \code{\link{owin}}.}
}
\value{
The converted polygon of class \code{"gpc.poly"} or \code{"owin"},
respectively. If neither package \pkg{rgeos} nor \pkg{gpclib} are available,
\code{owin2gpc} will just return the \code{pts} slot of the
\code{"gpc.poly"} (no formal class) with a warning.
}
\description{
Package \pkg{polyCub} implements converters between the classes
\code{"\link[spatstat:owin.object]{owin}"} of package \pkg{spatstat} and
\code{"\link[rgeos:gpc.poly-class]{gpc.poly}"} of package \pkg{rgeos}
(originally from \pkg{gpclib}).
Support for the \code{"gpc.poly"} class was dropped from
\pkg{spatstat} as of version 1.34-0.
}
\note{
The converter \code{owin2gpc} requires the package \pkg{rgeos} (or
\pkg{gpclib}) for the formal class definition of a \code{"gpc.poly"}.
It will produce vertices ordered according to the \pkg{sp} convention,
i.e. clockwise for normal boundaries and anticlockwise for holes, where,
however, the first vertex is \emph{not} repeated!
}
\author{
Sebastian Meyer
}
\seealso{
\code{\link{xylist}}, and the package \pkg{rgeos} for
conversions of \code{"gpc.poly"} objects from and to \pkg{sp}'s
\code{"\linkS4class{SpatialPolygons}"} class.
}
\keyword{methods}
\keyword{spatial}

