% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{isClosed}
\alias{isClosed}
\title{Check if Polygon is Closed}
\usage{
isClosed(coords)
}
\arguments{
\item{coords}{numeric coordinate matrix. It is interpreted by
\code{\link{xy.coords}}.}
}
\value{
logical
}
\description{
Check if the first and last coordinates of a coordinate matrix are
identical.
}
\keyword{internal}
\keyword{spatial}
