% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plot_polyregion}
\alias{plot_polyregion}
\title{Plots a Polygonal Domain (of Various Classes)}
\usage{
plot_polyregion(polyregion, lwd = 2, add = FALSE)
}
\arguments{
\item{polyregion}{a polygonal domain.
The following classes are supported:
\code{"\link[spatstat.geom]{owin}"} from package \pkg{spatstat.geom},
\code{"gpc.poly"} from \pkg{gpclib},
\code{"\link[sp:SpatialPolygons-class]{SpatialPolygons}"},
\code{"\link[sp:Polygons-class]{Polygons}"},
and \code{"\link[sp:Polygon-class]{Polygon}"} from package \pkg{sp}, as well as
\code{"\link[sf:st_polygon]{(MULTI)POLYGON}"} from package \pkg{sf}.
(For these classes, \pkg{polyCub} knows how to get an \code{\link{xylist}}.)}

\item{lwd}{line width of the polygon edges.}

\item{add}{logical. Add to existing plot?}
}
\description{
Plots a Polygonal Domain (of Various Classes)
}
