% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{polyCub-package}
\alias{polyCub-package}
\title{Cubature over Polygonal Domains}
\description{
The \R package \pkg{polyCub} \bibcitep{polyCub::CITATION} implements
\emph{cubature} (numerical integration) over \emph{polygonal} domains.
It solves the problem of integrating a continuously differentiable
function \eqn{f(x,y)} over simple closed polygons.
}
\details{
\pkg{polyCub} provides the following cubature methods:
\describe{
\item{\code{\link{polyCub.SV}}:}{
General-purpose \emph{product Gauss cubature}
\bibcitep{polyCub::sommariva.vianello2007}
}
\item{\code{\link{polyCub.midpoint}}:}{
Simple \emph{two-dimensional midpoint rule} based on
\code{\link[spatstat.geom]{as.im.function}} from \pkg{spatstat.geom}
(Baddeley et al. 2015)
}
\item{\code{\link{polyCub.iso}}:}{
Adaptive cubature for \emph{radially symmetric functions}
via line \code{\link{integrate}()} along the polygon boundary
\bibcitep{|polyCub::meyer.held2014|Supplement B\\\\\\, Section 2.4}
}
}
A brief description and benchmark experiment of the above cubature
methods can be found in the \code{vignette("polyCub")}.

There is also \code{\link{polyCub.exact.Gauss}}, intended to
accurately (but slowly) integrate the \emph{bivariate Gaussian density};
however, this implementation is disabled as of \pkg{polyCub} 0.9.0:
it needs a reliable implementation of polygon triangulation.

\bibcitet{|polyCub::meyer2010|Section 3.2}
discusses and compares some of these methods.
}
\references{
Baddeley A., Rubak E., Turner R. (2015).
\emph{Spatial Point Patterns: Methodology and Applications with R}.
Chapman and Hall/CRC Press, London.

\bibshow{*}
}
\seealso{
\code{vignette("polyCub")}

For the special case of a rectangular domain along the axes
(e.g., a bounding box), the \CRANpkg{cubature} package is more appropriate.
}
\author{
Sebastian Meyer
}
