% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{vecnorm}
\alias{vecnorm}
\title{Euclidean Vector Norm (Length)}
\usage{
vecnorm(x)
}
\arguments{
\item{x}{numeric vector.}
}
\value{
\code{sqrt(sum(x^2))}
}
\description{
This is nothing else than \code{sqrt(sum(x^2))}.
}
\keyword{internal}
\keyword{math}
