\name{coerce-sp-methods}
\alias{coerce,Polygon,Polygons-method}
\alias{coerce,Polygon,owin-method}
\alias{coerce,Polygons,owin-method}
\alias{coerce,SpatialPolygons,owin-method}
\alias{coerce-sp-methods}
\title{Coerce \code{"Polygons"} to \code{"owin"}}
\description{
Package \pkg{polyCub} also implements \code{coerce}-methods
(\code{as(object, Class)}) to convert
\code{"\link[sp:Polygons-class]{Polygons}"} (or
\code{"\link[sp:SpatialPolygons-class]{SpatialPolygons}"}
or \code{"\link[sp:Polygon-class]{Polygon}"}) to
\code{"\link[spatstat:owin.object]{owin}"}.
}
\author{
Sebastian Meyer
}
\keyword{methods}
\keyword{spatial}

