% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce-sp-methods.R
\name{coerce-sp-methods}
\alias{coerce-sp-methods}
\alias{as.owin.SpatialPolygons}
\alias{as.owin.Polygons}
\alias{as.owin.Polygon}
\alias{coerce,SpatialPolygons,owin-method}
\alias{coerce,Polygons,owin-method}
\alias{coerce,Polygon,owin-method}
\alias{coerce,Polygon,Polygons-method}
\title{Coerce \code{"SpatialPolygons"} to \code{"owin"}}
\usage{
as.owin.SpatialPolygons(W, ...)

as.owin.Polygons(W, ...)

as.owin.Polygon(W, ...)
}
\arguments{
\item{W}{an object of class \code{"SpatialPolygons"},
\code{"Polygons"}, or \code{"Polygon"}.}

\item{...}{further arguments passed to \code{\link[spatstat]{owin}}.}
}
\description{
Package \pkg{polyCub} implements \code{coerce}-methods
(\code{as(object, Class)}) to convert \code{"\linkS4class{SpatialPolygons}"}
(or \code{"\linkS4class{Polygons}"} or \code{"\linkS4class{Polygon}"})
to \code{"\link[spatstat:owin.object]{owin}"}.
They are also available as \code{as.owin.*} functions to support
\code{\link{polyCub.midpoint}}. However, these are no registered S3 methods
for \code{\link[spatstat]{as.owin}}, since package \pkg{spatstat} is
optional.
Note that the \pkg{maptools} package contains an alternative implementation
of coercion from \code{"SpatialPolygons"} to \code{"owin"} (and reverse),
and \R will use the S4 \code{coerce}-method that was loaded last,
and prefer the \code{as.owin.SpatialPolygons} S3-method exported from
\pkg{maptools} if attached.
}
\examples{
if (require("spatstat") && require("sp")) {
    diamond <- list(x = c(1,2,1,0), y = c(1,2,3,2))  # anti-clockwise
    diamond.owin <- owin(poly = diamond)
    diamond.sp <- Polygon(lapply(diamond, rev))      # clockwise
    diamond.owin_from_sp <- as(diamond.sp, "owin")
    stopifnot(all.equal(diamond.owin, diamond.owin_from_sp))

    ## similarly works for Polygons and SpatialPolygons
    diamond.Ps <- as(diamond.sp, "Polygons")
    stopifnot(identical(diamond.owin, as.owin(diamond.Ps)))
    diamond.SpPs <- SpatialPolygons(list(diamond.Ps))
    stopifnot(identical(diamond.owin, as.owin(diamond.SpPs)))
}
}
\author{
Sebastian Meyer
}
\keyword{methods}
\keyword{spatial}
