% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{dotprod}
\alias{dotprod}
\title{Dot/Scalar Product of Two Vectors}
\usage{
dotprod(x, y)
}
\arguments{
\item{x, y}{numeric vectors (of compatible lengths).}
}
\value{
\code{sum(x*y)}
}
\description{
This is nothing else than \code{sum(x*y)}.
}
\keyword{internal}
\keyword{math}
