% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{gpclibPermit}
\alias{gpclibPermit}
\alias{gpclibPermitStatus}
\title{\pkg{gpclib} License Acceptance (OBSOLETE)}
\usage{
gpclibPermit()

gpclibPermitStatus()
}
\description{
Previous versions of package \CRANpkg{gpclib} had a restricted license
(commercial use prohibited) and these functions were used as a blocker.
They now always return \code{TRUE}.
}
\details{
\pkg{gpclib} functionality is only required for
\code{\link{polyCub.exact.Gauss}}.
}
\keyword{internal}
