% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{gpclibPermit}
\alias{gpclibPermit}
\alias{gpclibPermitStatus}
\title{\pkg{gpclib} License Acceptance}
\usage{
gpclibPermit()

gpclibPermitStatus()
}
\description{
Similar to the handling in package \pkg{maptools}, these functions
explicitly accept the restricted \pkg{gpclib} license (commercial use
prohibited) and report its acceptance status, respectively.
\pkg{gpclib} functionality is only required for
\code{\link{polyCub.exact.Gauss}}.
}
