% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{polyCub-package}
\alias{polyCub-package}
\title{Cubature over Polygonal Domains}
\description{
The \R package \pkg{polyCub} implements \emph{cubature}
(numerical integration) over \emph{polygonal} domains.
It solves the problem of integrating a continuously differentiable
function \eqn{f(x,y)} over simple closed polygons.
}
\details{
\pkg{polyCub} provides the following cubature methods,
which can either be called explicitly or via the generic
\code{\link{polyCub}} function:
\describe{
\item{\code{\link{polyCub.SV}}:}{
General-purpose \emph{product Gauss cubature} (Sommariva and Vianello, 2007)
}
\item{\code{\link{polyCub.midpoint}}:}{
Simple \emph{two-dimensional midpoint rule} based on
\code{\link[spatstat]{as.im.function}} from \pkg{spatstat}
(Baddeley and Turner, 2005)
}
\item{\code{\link{polyCub.iso}}:}{
Adaptive cubature for \emph{radially symmetric functions}
via line \code{\link{integrate}()} along the polygon boundary
(Meyer and Held, 2014, Supplement B, Section 2.4).
}
\item{\code{\link{polyCub.exact.Gauss}}:}{
Accurate (but slow) integration of the \emph{bivariate Gaussian density}
based on polygon triangulation (via \code{\link[gpclib]{tristrip}} from
\pkg{gpclib}) and (numerous) evaluations of cumulative densities (via
\code{\link[mvtnorm]{pmvnorm}} from package \pkg{mvtnorm}).
Note that there is also a function \code{\link{circleCub.Gauss}}
to integrate the \emph{isotropic} Gaussian density over a
\emph{circular domain}.
}
}
A more detailed description and benchmark experiment of the above cubature
methods can be found in the \code{vignette("polyCub")} and in
Meyer (2010, Section 3.2).
}
\references{
Abramowitz, M. and Stegun, I. A. (1972).
Handbook of Mathematical Functions with Formulas, Graphs, and Mathematical
Tables. New York: Dover Publications.

Baddeley, A. and Turner, R. (2005).
\pkg{spatstat}: an \R package for analyzing spatial point patterns.
\emph{Journal of Statistical Software}, \bold{12} (6), 1-42.

Meyer, S. (2010).
Spatio-Temporal Infectious Disease Epidemiology based on Point Processes.
Master's Thesis, LMU Munich.
Available as \url{http://epub.ub.uni-muenchen.de/11703/}.

Meyer, S. and Held, L. (2014).
Power-law models for infectious disease spread.
\emph{The Annals of Applied Statistics}, \bold{8} (3), 1612-1639.\cr
DOI-Link: \url{https://doi.org/10.1214/14-AOAS743},
\href{https://arxiv.org/abs/1308.5115}{arXiv:1308.5115}

Sommariva, A. and Vianello, M. (2007).
Product Gauss cubature over polygons based on Green's integration formula.
\emph{BIT Numerical Mathematics}, \bold{47} (2), 441-453.\cr
DOI-Link: \url{https://doi.org/10.1007/s10543-007-0131-2}
}
\seealso{
\code{vignette("polyCub")}

For the special case of a rectangular domain along the axes
(e.g., a bounding box), the \pkg{cubature} package is more appropriate.
}
